--- src/components/Contact.js
@@ -1,28 +1,0 @@
-import React from 'react';
-
-function Contact() {
-  return (
-    <div className="contact-page">
-      <h2>You can contact us on:</h2>
-      <ul>
-        <li>Benson Njuguna</li>
-        <li>Fullstack Software Developer</li>
-        <li>Email: njugunab655@gmail.com</li>
-        <li>
-          GitHub:
-          <a href="https://github.com/Bennyjoez">
-            @bennyjoez
-          </a>
-        </li>
-        <li>
-          LinkedIn:
-          <a href="https://www.linkedin.com/in/bennyjoez/">
-            @bennyjoez
-          </a>
-        </li>
-      </ul>
-    </div>
-  );
-}
-
-export default Contact;
