+++ src/components/Profile.js
@@ -1,0 +1,34 @@
+import React from 'react';
+
+function Contact() {
+  return (
+    <div className="profile-page">
+      <div className="under-construction">
+        Page Under Construction...
+      </div>
+      <hr />
+      <div className="contact-page">
+        <h2>For inquiries contact us on:</h2>
+        <ul>
+          <li>Benson Njuguna</li>
+          <li>Fullstack Software Developer</li>
+          <li>Email: njugunab655@gmail.com</li>
+          <li>
+            GitHub:
+            <a href="https://github.com/Bennyjoez">
+              @bennyjoez
+            </a>
+          </li>
+          <li>
+            LinkedIn:
+            <a href="https://www.linkedin.com/in/bennyjoez/">
+              @bennyjoez
+            </a>
+          </li>
+        </ul>
+      </div>
+    </div>
+  );
+}
+
+export default Contact;
